Object Pooling Functional
All basic functionality of the object pooling system is in.  Pools are created by scriptable object. Then references to the pools you would like to generate are given to the Pool Manager Monobehavior.  Pools will be created on Awake().  Pools can expand to accomidate for a large demand for objects than was originally anticipated. Pools can consist of several different types of gameobjects.